# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/space/create": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "新建空间",
                "parameters": [
                    {
                        "description": "名称和描述",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateSpacePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"空间详情",
                        "schema": {
                            "$ref": "#/definitions/model.Space"
                        }
                    }
                }
            }
        },
        "/space/delete": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "删除空间",
                "parameters": [
                    {
                        "description": "空间ID列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DeleteSpacePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/space/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取空间详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "空间ID",
                        "name": "space_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"空间详情",
                        "schema": {
                            "$ref": "#/definitions/entity.GetSpaceInfoResp"
                        }
                    }
                }
            }
        },
        "/space/search": {
            "get": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "搜索空间",
                "parameters": [
                    {
                        "type": "string",
                        "name": "keywords",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "owner_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"空间列表",
                        "schema": {
                            "$ref": "#/definitions/entity.GetSpacesResp"
                        }
                    }
                }
            }
        },
        "/space/update": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "更新空间",
                "parameters": [
                    {
                        "description": "空间ID、名称和描述",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateSpacePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"更新后的空间详情",
                        "schema": {
                            "$ref": "#/definitions/model.Space"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.CreateSpacePayload": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entity.DeleteSpacePayload": {
            "type": "object",
            "required": [
                "space_ids"
            ],
            "properties": {
                "space_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "entity.GetSpaceInfoResp": {
            "type": "object",
            "properties": {
                "creation_time": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "space_id": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
        "entity.GetSpacesResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Space"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "entity.UpdateSpacePayload": {
            "type": "object",
            "required": [
                "space_id"
            ],
            "properties": {
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "space_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Space": {
            "type": "object",
            "properties": {
                "creation_time": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "space_id": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
...
    }
}